finally fixed deleting messages

+++ app/assets/javascripts/bundle.js
@@ -24530,7 +24530,7 @@
 	      }
 	    }).bind(this));
 	    ApiUserUtil.updateProfile(user, (function (id) {
-	      this.history.pushState(null, "/user/" + id, {});
+	      this.history.pushState(null, "/", {});
 	    }).bind(this));
 	    this.setState(this.blankAttrs);
 	  },
@@ -33315,7 +33315,7 @@
 	    var messages = MessageStore.allMyReceivedMessages(current_user_id);
 	    var count = 0;
 	    messages.forEach(function (message) {
-	      if (!message.read) {
+	      if (!message.read && !message.receiver_delete) {
 	        count += 1;
 	      }
 	    });
@@ -33350,6 +33350,19 @@
 	  showMessage: function (event) {
 	    event.preventDefault();
 	    this.setState({ messageDetails: parseInt(event.currentTarget.id) });
+	    var message = MessageStore.find(parseInt(event.currentTarget.id));
+	    if (message.receiver_id === this.state.current_user_id && !message.read) {
+	      ApiMessageUtil.updateMessage({
+	        body: message.body,
+	        sender_id: message.sender_id,
+	        receiver_id: message.receiver_id,
+	        read: true,
+	        id: message.id,
+	        created_at: message.created_at,
+	        receiver_delete: message.receiver_delete,
+	        sender_delete: message.sender_delete
+	      });
+	    }
 	  },
 	
 	  deleteMessage: function (event) {
@@ -33359,7 +33372,31 @@
 	      this.setState({ messageDetails: undefined });
 	    }
 	    var message = MessageStore.find(id);
-	    ApiMessageUtil.deleteMessage(message);
+	    if (message.sender_id === this.state.current_user_id && !message.receiver_delete) {
+	      ApiMessageUtil.updateMessage({
+	        body: message.body,
+	        sender_id: message.sender_id,
+	        receiver_id: message.receiver_id,
+	        read: message.read,
+	        id: message.id,
+	        created_at: message.created_at,
+	        receiver_delete: message.receiver_delete,
+	        sender_delete: true
+	      });
+	    } else if (message.receiver_id === this.state.current_user_id && !message.sender_delete) {
+	      ApiMessageUtil.updateMessage({
+	        body: message.body,
+	        sender_id: message.sender_id,
+	        receiver_id: message.receiver_id,
+	        read: message.read,
+	        id: message.id,
+	        created_at: message.created_at,
+	        receiver_delete: true,
+	        sender_delete: message.sender_delete
+	      });
+	    } else {
+	      ApiMessageUtil.deleteMessage(message);
+	    }
 	  },
 	
 	  render: function () {
@@ -33381,29 +33418,31 @@
 	          } else {
 	            var read = "unread";
 	          }
-	          messageReceivedContainer.unshift(React.createElement(
-	            'div',
-	            { key: message.id },
-	            React.createElement(
-	              'button',
-	              { title: 'delete', className: 'remove-interest', id: message.id, onClick: this.deleteMessage },
-	              'X'
-	            ),
-	            React.createElement(
-	              'li',
-	              { onClick: this.showMessage, id: message.id, className: 'message-list-item' },
+	          if (!message.receiver_delete) {
+	            messageReceivedContainer.unshift(React.createElement(
+	              'div',
+	              { key: message.id },
 	              React.createElement(
-	                'div',
-	                { className: read },
-	                user.username,
+	                'button',
+	                { title: 'delete', className: 'remove-interest', id: message.id, onClick: this.deleteMessage },
+	                'X'
+	              ),
+	              React.createElement(
+	                'li',
+	                { onClick: this.showMessage, id: message.id, className: 'message-list-item' },
 	                React.createElement(
-	                  'span',
-	                  { className: 'date' },
-	                  date
+	                  'div',
+	                  { className: read },
+	                  user.username,
+	                  React.createElement(
+	                    'span',
+	                    { className: 'date' },
+	                    date
+	                  )
 	                )
 	              )
-	            )
-	          ));
+	            ));
+	          }
 	        }
 	      }).bind(this));
 	      messageReceivedContainer.unshift(React.createElement(
@@ -33429,29 +33468,31 @@
 	        var user = UserStore.find(parseInt(message.receiver_id));
 	        var date = MessageStore.dateToString(message.created_at);
 	        if (user) {
-	          messageSentContainer.unshift(React.createElement(
-	            'div',
-	            { key: message.id },
-	            React.createElement(
-	              'button',
-	              { title: 'delete', className: 'remove-interest', id: message.id, onClick: this.deleteMessage },
-	              'X'
-	            ),
-	            React.createElement(
-	              'li',
-	              { onClick: this.showMessage, id: message.id, className: 'message-list-item' },
+	          if (!message.sender_delete) {
+	            messageSentContainer.unshift(React.createElement(
+	              'div',
+	              { key: message.id },
 	              React.createElement(
-	                'div',
-	                { className: 'read' },
-	                user.username,
+	                'button',
+	                { title: 'delete', className: 'remove-interest', id: message.id, onClick: this.deleteMessage },
+	                'X'
+	              ),
+	              React.createElement(
+	                'li',
+	                { onClick: this.showMessage, id: message.id, className: 'message-list-item' },
 	                React.createElement(
-	                  'span',
-	                  { className: 'date' },
-	                  date
+	                  'div',
+	                  { className: 'read' },
+	                  user.username,
+	                  React.createElement(
+	                    'span',
+	                    { className: 'date' },
+	                    date
+	                  )
 	                )
 	              )
-	            )
-	          ));
+	            ));
+	          }
 	        }
 	      }).bind(this));
 	      messageSentContainer.unshift(React.createElement(
@@ -33586,16 +33627,6 @@
 	        receiver: UserStore.find(parseInt(message.receiver_id)),
 	        created_at: message.created_at
 	      });
-	      if (message.receiver_id === this.state.current_user_id && !message.read) {
-	        ApiMessageUtil.updateMessage({
-	          body: message.body,
-	          sender_id: message.sender_id,
-	          receiver_id: message.receiver_id,
-	          read: true,
-	          id: message.id,
-	          created_at: message.created_at
-	        });
-	      }
 	    }
 	  },
 	
@@ -33807,7 +33838,7 @@
 	    var myMessages = MessageStore.allMyReceivedMessages(this.state.current_user.id);
 	    var unread_count = 0;
 	    myMessages.forEach(function (message) {
-	      if (!message.read) {
+	      if (!message.read && !message.receiver_delete) {
 	        unread_count += 1;
 	      }
 	    });
